- [OBMFULL-3155] Cleaning and refactoring

+++ java/sync/opush/push-bean/src/test/java/org/obm/push/bean/MSEventToObmEventConverterTest.java
@@ -1,3 +1,34 @@
+/* ***** BEGIN LICENSE BLOCK *****
+ * 
+ * Copyright (C) 2011-2012  Linagora
+ *
+ * This program is free software: you can redistribute it and/or 
+ * modify it under the terms of the GNU Affero General Public License as 
+ * published by the Free Software Foundation, either version 3 of the 
+ * License, or (at your option) any later version, provided you comply 
+ * with the Additional Terms applicable for OBM connector by Linagora 
+ * pursuant to Section 7 of the GNU Affero General Public License, 
+ * subsections (b), (c), and (e), pursuant to which you must notably (i) retain 
+ * the “Message sent thanks to OBM, Free Communication by Linagora” 
+ * signature notice appended to any and all outbound messages 
+ * (notably e-mail and meeting requests), (ii) retain all hypertext links between 
+ * OBM and obm.org, as well as between Linagora and linagora.com, and (iii) refrain 
+ * from infringing Linagora intellectual property rights over its trademarks 
+ * and commercial brands. Other Additional Terms apply, 
+ * see <http://www.linagora.com/licenses/> for more details. 
+ *
+ * This program is distributed in the hope that it will be useful, 
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY 
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License 
+ * for more details. 
+ *
+ * You should have received a copy of the GNU Affero General Public License 
+ * and its applicable Additional Terms for OBM along with this program. If not, 
+ * see <http://www.gnu.org/licenses/> for the GNU Affero General Public License version 3 
+ * and <http://www.linagora.com/licenses/> for the Additional Terms applicable to 
+ * OBM connectors. 
+ * 
+ * ***** END LICENSE BLOCK ***** */
 package org.obm.push.bean;
 
 import java.util.ArrayList;
@@ -418,7 +449,7 @@ public class MSEventToObmEventConverterTest {
 		
 		Event convertedEvent = convertToOBMEvent(msEvent);
 		
-		int reminderInSecond = (int) minuteToSecond(msEvent.getReminder());
+		int reminderInSecond = minuteToSecond(msEvent.getReminder());
 		Assertions.assertThat(convertedEvent.getAlert()).isEqualTo(reminderInSecond);
 	}
 
@@ -760,7 +791,7 @@ public class MSEventToObmEventConverterTest {
 
 		Assertions.assertThat(converted.getStartDate()).isEqualTo(msEvent.getStartTime());
 		Assertions.assertThat(converted.getEndDate()).isEqualTo(msEvent.getEndTime());
-		Assertions.assertThat(converted.getDuration()).isEqualTo((int) getOneYearInSecond());
+		Assertions.assertThat(converted.getDuration()).isEqualTo(getOneYearInSecond());
 	}
 
 	@Test
@@ -869,15 +900,15 @@ public class MSEventToObmEventConverterTest {
 		return org.obm.DateUtils.date(date);
 	}
 
-	private long getOneDayInSecond() {
-		return DateUtils.daysToSeconds(1);
+	private int getOneDayInSecond() {
+		return (int) DateUtils.daysToSeconds(1);
 	}
 
-	private long getOneYearInSecond() {
-		return DateUtils.yearsToSeconds(1);
+	private int getOneYearInSecond() {
+		return (int) DateUtils.yearsToSeconds(1);
 	}
 
-	private long minuteToSecond(int minutes) {
-		return DateUtils.minutesToSeconds(minutes);
+	private int minuteToSecond(int minutes) {
+		return (int) DateUtils.minutesToSeconds(minutes);
 	}
 }
